Revert "[web] Switch to doubles (#37336)" (flutter/engine#37609)
This reverts commit 1a0140ccea32c88cc69d5a65e1cf229b75165516.

+++ engine/src/flutter/lib/web_ui/lib/src/engine/assets.dart
@@ -84,7 +84,7 @@ class AssetManager {
           printWarning('Asset manifest does not exist at `$url` – ignoring.');
           return Uint8List.fromList(utf8.encode('{}')).buffer.asByteData();
         }
-        throw AssetManagerException(url, request.status!.toInt());
+        throw AssetManagerException(url, request.status!);
       }
 
       final String? constructorName = target == null ? 'null' :
